+++ MinhaLeitura/SRC/Controller/CadastroController.java
@@ -1,0 +1,13 @@
+package SRC.Controller;
+
+import javafx.fxml.Initializable;
+
+import java.net.URL;
+import java.util.ResourceBundle;
+
+public class CadastroController implements Initializable {
+    @Override
+    public void initialize(URL url, ResourceBundle resourceBundle) {
+
+    }
+}
