+++ MinhaLeitura/SRC/Controller/EmLeituraController.java
@@ -1,0 +1,53 @@
+package SRC.Controller;
+
+import SRC.Model.VO.User;
+import SRC.View.Telas;
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
+import javafx.fxml.Initializable;
+import javafx.scene.control.Label;
+
+import java.net.URL;
+import java.util.ResourceBundle;
+
+public class EmLeituraController implements Initializable {
+
+    private User usuario;
+    @FXML
+    public Label nomeUsuario;
+
+    @Override
+    public void initialize(URL url, ResourceBundle resourceBundle) {
+       nomeUsuario.setText(usuario.getUsername());
+    }
+
+    public EmLeituraController() {
+
+        usuario = FrontController.getUsuario();
+        System.out.println(usuario.getUsername());
+    }
+
+    public void sair(ActionEvent event) throws Exception{
+        Telas.telaLogin();
+    }
+
+    public void telaInicial(ActionEvent event) throws  Exception{
+        Telas.telaInicial(usuario);
+    }
+
+    public void telaLeituras(ActionEvent event) throws  Exception{
+        Telas.telaLeituras();
+    }
+
+    public void telaLeiturasFinalizadas(ActionEvent event) throws  Exception{
+        Telas.telaLeiturasFinalizadas();
+    }
+
+    public void telaConfiguracoes(ActionEvent event) throws  Exception{
+        Telas.telaConfiguracoes();
+    }
+
+    public void abrirModalAddLeitura(ActionEvent event) throws  Exception{
+        Telas.modalAddLeitura();
+    }
+}
