+++ MinhaLeitura/SRC/View/VE/Configuracoes.fxml
@@ -1,17 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import java.lang.String?>
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.Cursor?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.Pane?>
-<?import javafx.scene.text.Font?>
+<?import java.lang.*?>
+<?import javafx.geometry.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.image.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.text.*?>
 
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="720.0" prefWidth="1020.0" style="-fx-background-color: #DDDDDA;" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="SRC.Controller.FrontController">
+<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="720.0" prefWidth="1020.0" style="-fx-background-color: #DDDDDA;" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="SRC.Controller.ConfiguracoesController">
    <children>
       <ImageView fitHeight="681.0" fitWidth="1001.0" layoutX="9.5" layoutY="19.5" pickOnBounds="true" preserveRatio="true">
          <image>
@@ -81,7 +78,7 @@
             <Image url="@../Images/Group%205%201.png" />
          </image>
       </ImageView>
-      <Label layoutX="536.0" layoutY="191.0" text="Nome Usuário">
+      <Label fx:id="nomeUsuario" layoutX="536.0" layoutY="191.0" text="Nome Usuário">
          <font>
             <Font name="System Bold" size="30.0" />
          </font>
