+++ MinhaLeitura/SRC/View/VE/EmLeitura.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.*?>
 
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="720.0" prefWidth="1020.0" style="-fx-background-color: #DDDDDA;" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="SRC.Controller.FrontController">
+<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="720.0" prefWidth="1020.0" style="-fx-background-color: #DDDDDA;" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="SRC.Controller.EmLeituraController">
    <children>
       <ImageView fitHeight="681.0" fitWidth="1001.0" layoutX="9.5" layoutY="19.5" pickOnBounds="true" preserveRatio="true">
          <image>
@@ -86,7 +86,7 @@
             <Cursor fx:constant="HAND" />
          </cursor>
       </Button>
-      <Label layoutX="783.0" layoutY="41.0" prefHeight="33.0" prefWidth="141.0" text="Nome teste" wrapText="true">
+      <Label fx:id="nomeUsuario" layoutX="783.0" layoutY="41.0" prefHeight="33.0" prefWidth="141.0" text="Nome teste" wrapText="true">
          <font>
             <Font name="System Bold" size="20.0" />
          </font>
