+++ MinhaLeitura/SRC/View/VE/Home.fxml
@@ -1,14 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import java.lang.*?>
-<?import javafx.geometry.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.image.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
+<?import java.lang.String?>
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.Cursor?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.ListView?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.Pane?>
+<?import javafx.scene.text.Font?>
 
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="720.0" prefWidth="1020.0" style="-fx-background-color: #DDDDDA;" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="SRC.Controller.HomeController">
+<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="720.0" prefWidth="1020.0" style="-fx-background-color: #DDDDDA;" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="SRC.Controller.HomeController">
    <children>
       <ImageView fitHeight="681.0" fitWidth="1001.0" layoutX="9.5" layoutY="19.5" pickOnBounds="true" preserveRatio="true">
          <image>
@@ -109,7 +113,7 @@
             <Cursor fx:constant="TEXT" />
          </cursor>
       </TextField>
-      <ListView fx:id="minhasLeituras" layoutX="266.0" layoutY="163.0" orientation="HORIZONTAL" prefHeight="215.0" prefWidth="490.0" style="-fx-background-color: none;">
+      <ListView fx:id="minhasLeiturasAtuais" editable="true" layoutX="266.0" layoutY="163.0" orientation="HORIZONTAL" prefHeight="215.0" prefWidth="490.0" style="-fx-background-color: none;">
          <cursor>
             <Cursor fx:constant="HAND" />
          </cursor>
